Crate rotor

The mio-based framework for doing I/O in simple and composable way

More documentation in the guide

Modules

mio

void

Macros

rotor_compose!
Compose multiple state machines into single type

Structs

Config
Event loop configuration

EarlyScope
This is a structure that works similarly to Scope, but doesn't have a context

EventSet

LoopCreator
An object that is used to construct a loop

LoopInstance
Second stage of loop creation

Notifier
The object used to wakeup unrelated state machine

PollOpt

Response
The response of a state machine to the (mio) action

Scope
The structure passed to every action handler

Time
The current time

Timeout

TimerError

Enums

Compose2
Composes two state machines

SpawnError
Error when spawning a new state machine

Void
The empty type for cases which can't occur.

WakeupError
Error when waking up a connection

Traits

Evented
A value that may be registered with an EventLoop

GenericScope
A common part of Scope and EarlyScope

Machine
A trait that every state machine in the loop must implement