Struct rotor::mio::tcp::TcpListener

pub struct TcpListener { /* fields omitted */ }

Methods

impl TcpListener

fn bind(addr: &SocketAddr) -> Result<TcpListener, Error>

Convenience method to bind a new TCP listener to the specified address to receive new connections.

This function will take the following steps:

Create a new TCP socket.

Set the SO_REUSEADDR option on the socket.

Bind the socket to the specified address.

Call listen on the socket to prepare it to receive new connections.

If fine-grained control over the binding and listening process for a socket is desired then the net2::TcpBuilder methods can be used in combination with the TcpListener::from_listener method to transfer ownership into mio.

fn from_listener(listener: TcpListener, addr: &SocketAddr) -> Result<TcpListener, Error>

Creates a new TcpListener from an instance of a std::net::TcpListener type.

This function will set the listener provided into nonblocking mode on Unix, and otherwise the stream will just be wrapped up in an mio stream ready to accept new connections and become associated with an event loop.

The address provided must be the address that the listener is bound to.

fn accept(&self) -> Result<(TcpStream, SocketAddr), Error>

Accepts a new TcpStream.

Returns a Ok(None) when the socket WOULDBLOCK, this means the stream will be ready at a later point. If an accepted stream is returned, the address of the peer is returned along with it

fn local_addr(&self) -> Result<SocketAddr, Error>

Returns the local socket address of this listener.

fn try_clone(&self) -> Result<TcpListener, Error>

Creates a new independently owned handle to the underlying socket.

The returned TcpListener is a reference to the same socket that this object references. Both handles can be used to accept incoming connections and options set on one listener will affect the other.

fn set_ttl(&self, ttl: u32) -> Result<(), Error>

Sets the value for the IP_TTL option on this socket.

This value sets the time-to-live field that is used in every packet sent from this socket.

fn ttl(&self) -> Result<u32, Error>

Gets the value of the IP_TTL option for this socket.

For more information about this option, see set_ttl.

fn set_only_v6(&self, only_v6: bool) -> Result<(), Error>

Sets the value for the IPV6_V6ONLY option on this socket.

If this is set to true then the socket is restricted to sending and receiving IPv6 packets only. In this case two IPv4 and IPv6 applications can bind the same port at the same time.

If this is set to false then the socket can be used to send and receive packets from an IPv4-mapped IPv6 address.

fn only_v6(&self) -> Result<bool, Error>

Gets the value of the IPV6_V6ONLY option for this socket.

For more information about this option, see set_only_v6.

fn take_error(&self) -> Result<Option<Error>, Error>

Get the value of the SO_ERROR option on this socket.

This will retrieve the stored error in the underlying socket, clearing the field in the process. This can be useful for checking errors between calls.