Struct rotor::mio::tcp::TcpStream

pub struct TcpStream { /* fields omitted */ }

Methods

impl TcpStream

fn connect(addr: &SocketAddr) -> Result<TcpStream, Error>

Create a new TCP stream and issue a non-blocking connect to the specified address.

This convenience method is available and uses the system's default options when creating a socket which is then connected. If fine-grained control over the creation of the socket is desired, you can use net2::TcpBuilder to configure a socket and then pass its socket to TcpStream::connect_stream to transfer ownership into mio and schedule the connect operation.

fn connect_stream(stream: TcpStream, addr: &SocketAddr) -> Result<TcpStream, Error>

Creates a new TcpStream from the pending socket inside the given std::net::TcpBuilder, connecting it to the address specified.

This constructor allows configuring the socket before it's actually connected, and this function will transfer ownership to the returned TcpStream if successful. An unconnected TcpStream can be created with the net2::TcpBuilder type (and also configured via that route).

The platform specific behavior of this function looks like:

On Unix, the socket is placed into nonblocking mode and then a connect call is issued.

On Windows, the address is stored internally and the connect operation is issued when the returned TcpStream is registered with an event loop. Note that on Windows you must bind a socket before it can be connected, so if a custom TcpBuilder is used it should be bound (perhaps to INADDR_ANY) before this method is called.

fn peer_addr(&self) -> Result<SocketAddr, Error>

Returns the socket address of the remote peer of this TCP connection.

fn local_addr(&self) -> Result<SocketAddr, Error>

Returns the socket address of the local half of this TCP connection.

fn try_clone(&self) -> Result<TcpStream, Error>

Creates a new independently owned handle to the underlying socket.

The returned TcpStream is a reference to the same stream that this object references. Both handles will read and write the same stream of data, and options set on one stream will be propagated to the other stream.

fn shutdown(&self, how: Shutdown) -> Result<(), Error>

Shuts down the read, write, or both halves of this connection.

This function will cause all pending and future I/O on the specified portions to return immediately with an appropriate value (see the documentation of Shutdown).

fn set_nodelay(&self, nodelay: bool) -> Result<(), Error>

Sets the value of the TCP_NODELAY option on this socket.

If set, this option disables the Nagle algorithm. This means that segments are always sent as soon as possible, even if there is only a small amount of data. When not set, data is buffered until there is a sufficient amount to send out, thereby avoiding the frequent sending of small packets.

fn nodelay(&self) -> Result<bool, Error>

Gets the value of the TCP_NODELAY option on this socket.

For more information about this option, see set_nodelay.

fn set_keepalive_ms(&self, keepalive: Option<u32>) -> Result<(), Error>

Sets whether keepalive messages are enabled to be sent on this socket.

On Unix, this option will set the SO_KEEPALIVE as well as the TCP_KEEPALIVE or TCP_KEEPIDLE option (depending on your platform). On Windows, this will set the SIO_KEEPALIVE_VALS option.

If None is specified then keepalive messages are disabled, otherwise the number of milliseconds specified will be the time to remain idle before sending a TCP keepalive probe.

Some platforms specify this value in seconds, so sub-second millisecond specifications may be omitted.

fn keepalive_ms(&self) -> Result<Option<u32>, Error>

Returns whether keepalive messages are enabled on this socket, and if so the amount of milliseconds between them.

For more information about this option, see set_keepalive_ms.

fn set_ttl(&self, ttl: u32) -> Result<(), Error>

Sets the value for the IP_TTL option on this socket.

This value sets the time-to-live field that is used in every packet sent from this socket.

fn ttl(&self) -> Result<u32, Error>

Gets the value of the IP_TTL option for this socket.

For more information about this option, see set_ttl.

fn take_error(&self) -> Result<Option<Error>, Error>

Get the value of the SO_ERROR option on this socket.

This will retrieve the stored error in the underlying socket, clearing the field in the process. This can be useful for checking errors between calls.