Struct rotor::Response

pub struct Response<M, N>(_, _);

The response of a state machine to the (mio) action

This value is returned by many methods of the Machine trait.

Methods

impl<M: Sized, N: Sized> Response<M, N>

fn ok(machine: M) -> Response<M, N>

fn spawn(machine: M, result: N) -> Response<M, N>

fn done() -> Response<M, N>

fn map<T, U, S, R>(self, self_mapper: S, result_mapper: R) -> Response<T, U> where S: FnOnce(M) -> T, R: FnOnce(N) -> U